# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domains.Link:
    properties:
      campaignId:
        type: string
      created_at:
        type: string
      id:
        type: string
      productId:
        type: string
      short_code:
        type: string
      target_url:
        type: string
      updated_at:
        type: string
    type: object
  domains.Offer:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_checked_at:
        type: string
      marketplace:
        type: string
      price:
        type: number
      productId:
        type: string
      store_name:
        type: string
      updated_at:
        type: string
    type: object
  domains.Product:
    properties:
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      source_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.CreateLinkRequest:
    properties:
      campaign_id:
        type: string
      product_id:
        type: string
    required:
    - campaign_id
    - product_id
    type: object
  dto.CreateProductRequest:
    properties:
      marketplace:
        enum:
        - shopee
        - lazada
        type: string
      source_url:
        type: string
    required:
    - marketplace
    - source_url
    type: object
...
  dto.EmptyResponse:
    properties:
      code:
        example: 0
        type: integer
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
  dto.LinkResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/domains.Link'
      message:
        example: Link created successfully
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
  dto.LinksResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          $ref: '#/definitions/domains.Link'
        type: array
      message:
        example: Links retrieved successfully
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
...
  dto.OfferResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/domains.Offer'
      message:
        example: Offer retrieved successfully
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
  dto.ProductResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/domains.Product'
      message:
        example: Product retrieved successfully
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
  dto.ProductsResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          $ref: '#/definitions/domains.Product'
        type: array
      message:
        example: Products retrieved successfully
        type: string
      success:
        example: true
        type: boolean
      txn_id:
        example: txn_123456
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing marketplace affiliate links (Lazada, Shopee)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Market Place Affiliate API
  version: "1.0"
paths:
...
  /link:
    post:
      consumes:
      - application/json
      description: Create a new affiliate link for a product and campaign
      parameters:
      - description: Link request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LinkResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      security:
      - BearerAuth: []
      summary: Create affiliate link
      tags:
      - link
  /link/{link_id}:
    delete:
      description: Delete a link and all associated clicks
      parameters:
      - description: Link ID
        in: path
        name: link_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      security:
      - BearerAuth: []
      summary: Delete link
      tags:
      - link
    get:
      description: Get a specific link by its ID
      parameters:
      - description: Link ID
        in: path
        name: link_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LinkResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      summary: Get link by ID
      tags:
      - link
  /link/campaign/{campaignId}:
    get:
      description: Get all links associated with a campaign
      parameters:
      - description: Campaign ID
        in: path
        name: campaignId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LinksResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      summary: Get links by campaign
      tags:
      - link
  /link/redirect/{short_code}:
    get:
      description: Track click and redirect to the marketplace affiliate link
      parameters:
      - description: Short code
        in: path
        name: short_code
        required: true
        type: string
      responses:
        "302":
          description: Redirect to affiliate URL
          schema:
            type: string
      summary: Redirect to affiliate link
      tags:
      - link
  /link/short-code/{short_code}:
    get:
      description: Get a link by its short code
      parameters:
      - description: Short code
        in: path
        name: short_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LinkResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      summary: Get link by short code
      tags:
      - link
  /product:
    get:
      description: Get all products for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductsResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user products
      tags:
      - product
    post:
      consumes:
      - application/json
      description: Create a new affiliate product from marketplace URL
      parameters:
      - description: Product request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductsResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add a new product
      tags:
      - product
  /product/{productId}:
    delete:
      description: Delete a product and all associated links and clicks
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - product
    get:
      description: Get a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      summary: Get product by ID
      tags:
      - product
  /product/{productId}/offer:
    get:
      description: Get marketplace offers for a specific product
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OfferResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.EmptyResponse'
      security:
      - BearerAuth: []
      summary: Get product offers
      tags:
      - product
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"